samsung vs apple brand reviews(DEV project)/requirements.txt
praw>=7.0.0
python-dotenv>=1.0.0
